`Flags.VALUE` should return `Byte`

+++ src/commonMain/kotlin/cloud.drakon.tempest/util/Flags.kt
@@ -1,6 +1,6 @@
 package cloud.drakon.tempest.util
 
-enum class Flags(val VALUE: Short) {
+enum class Flags(val VALUE: Byte) {
     /** Do not include any embeds when serializing this message. */
     SUPPRESS_EMBEDS(1 shl 2),
